utils: add disable_output helper
This helper is useful for tools that support json output and want to
supress other kind of output without adding a lot of conditionals
everywhere. The function returns the original reference to sys.stdout
and a stub function to replace bcc.utils.printb().

If the tool uses printb, it should assign the returned stub to disable it:

stdout, printb = disable_stdout()

If the tool only uses print, then the stub can be ignored:

stdout, _ = disable_stdout()

The original reference to sys.stdout is saved on stdout and can be used
later on to print to stdout:

stdout.write("hi there...")