docker: Use debian bookworm as final image.
golang:1.19 was modified to use debian:bookworm as its base as bookworm is now
the new stable [1, 2].
bookworm comes with GLIBC 2.36 which is indeed not compatible with GLIBC 2.31
shipped in bullseye [3, 4].
This leaded to impossibility of running the gadgettracermanager compiled on
golang:1.19, so with GLIBC 2.36, and run in iovisor/bcc, i.e. ubuntu:focal which
comes with GLIBC 2.31 too [5].

To solve this, this commit uses a different final image for iovisor/bcc.
The debian packages will be built in ubuntu:focal but we will install them in
debian:bookworm to ensure we have the GLIBC between build time and run time.
We also removed everything related to python2 as it is no more developed.

Note that, for unknown reason, we needed to install libtinfo5 because libbcc.so
needs it.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://www.debian.org/News/2023/20230610
[2]: https://github.com/docker-library/golang/commit/cd607138e461
[3]: https://packages.debian.org/source/bookworm/glibc
[4]: https://packages.debian.org/source/bullseye/glibc
[5]: https://packages.ubuntu.com/source/focal/glibc